# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI-Ollama-RAG-App", "version": "0.1.0" },
  "paths": {
    "/api/v1/documents/ingest": {
      "post": {
        "tags": ["Documents"],
        "summary": "Ingest Document",
        "description": "End-to-end RAG ingestion pipeline:\n1. Hash PDF for deduplication.\n2. Extract text (PyMuPDF).\n3. Semantic chunking.\n4. Generate embeddings (Ollama).\n5. Bulk upsert into Neon (pgvector).",
        "operationId": "ingest_document_api_v1_documents_ingest_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_document_api_v1_documents_ingest_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IngestResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/documents/": {
      "get": {
        "tags": ["Documents"],
        "summary": "List User Files",
        "description": "Returns a list of all PDFs uploaded by the authenticated user.",
        "operationId": "list_user_files_api_v1_documents__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/documents/{file_id}": {
      "delete": {
        "tags": ["Documents"],
        "summary": "Delete User File",
        "description": "Deletes a specific file and all of its associated vector chunks.",
        "operationId": "delete_user_file_api_v1_documents__file_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "File Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/search": {
      "post": {
        "tags": ["Chat"],
        "summary": "Raw Vector Search",
        "description": "Directly queries the vector database using semantic similarity.\nReturns the raw chunks and cosine similarity scores without LLM generation.\nOnly searches the authenticated user's documents.",
        "operationId": "raw_vector_search_api_v1_chat_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/SearchResult" },
                  "type": "array",
                  "title": "Response Raw Vector Search Api V1 Chat Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/chat/completions": {
      "post": {
        "tags": ["Chat"],
        "summary": "Chat Completions",
        "description": "End-to-end RAG Chat Endpoint.\n1. Embeds the user's query.\n2. Retrieves the top matching chunks from the database (isolated to user).\n3. Streams the LLM's grounded response back to the client.",
        "operationId": "chat_completions_api_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/auth/request-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request Otp",
        "description": "Initiates the registration flow by generating and emailing a 6-digit OTP.\nFails silently if the email is already registered to prevent enumeration.",
        "operationId": "request_otp_api_v1_auth_request_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OTPRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register",
        "description": "Verifies the OTP and registers a new user with a hashed password.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "description": "Authenticates a user and returns a JSON Web Token (JWT).\nNote: OAuth2 requires the field to be named 'username', which mapped to email.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TokenResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/request-reset-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request Reset Otp",
        "description": "Initiates the password reset flow by sending an OTP.\nFails silently if the email is not registered.",
        "operationId": "request_reset_otp_api_v1_auth_request_reset_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OTPRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Reset Password",
        "description": "Verifies the OTP and updates the user's password.",
        "operationId": "reset_password_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordResetRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_ingest_document_api_v1_documents_ingest_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_ingest_document_api_v1_documents_ingest_post"
      },
      "Body_login_api_v1_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              { "type": "string", "pattern": "^password$" },
              { "type": "null" }
            ],
            "title": "Grant Type"
          },
          "username": { "type": "string", "title": "Username" },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": { "type": "string", "title": "Scope", "default": "" },
          "client_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "Body_login_api_v1_auth_login_post"
      },
      "ChatRequest": {
        "properties": {
          "query": { "type": "string", "title": "Query" },
          "limit": { "type": "integer", "title": "Limit", "default": 5 }
        },
        "type": "object",
        "required": ["query"],
        "title": "ChatRequest",
        "description": "Represents a chat payload."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "status": { "type": "string", "title": "Status" },
          "filename": { "type": "string", "title": "Filename" },
          "chunks_processed": {
            "type": "integer",
            "title": "Chunks Processed"
          },
          "detail": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["status", "filename", "chunks_processed"],
        "title": "IngestResponse"
      },
      "OTPRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" }
        },
        "type": "object",
        "required": ["email"],
        "title": "OTPRequest"
      },
      "PasswordResetRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "otp": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Otp"
          },
          "new_password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "New Password",
            "description": "Must be between 8 and 256 characters"
          }
        },
        "type": "object",
        "required": ["email", "otp", "new_password"],
        "title": "PasswordResetRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": { "type": "string", "format": "email", "title": "Email" },
          "otp": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Otp"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "description": "Must be between 8 and 256 characters"
          }
        },
        "type": "object",
        "required": ["email", "otp", "password"],
        "title": "RegisterRequest"
      },
      "SearchResult": {
        "properties": {
          "text": { "type": "string", "title": "Text" },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "similarity_score": {
            "type": "number",
            "title": "Similarity Score",
            "description": "Cosine similarity score (1.0 is exact match)"
          }
        },
        "type": "object",
        "required": ["text", "metadata", "similarity_score"],
        "title": "SearchResult",
        "description": "Represents a chunk of text retrieved from the vector database,\nincluding its relevance score."
      },
      "TokenResponse": {
        "properties": {
          "access_token": { "type": "string", "title": "Access Token" },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": ["access_token"],
        "title": "TokenResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/auth/login" }
        }
      }
    }
  }
}